# OCPBUGS-32469: Remove tuned/rendered object