… pkg/operator/profilecalculator_test.go
… pkg/tuned/controller.go