The NTO operand is controlled by the operator by updates to two resources.  Its corresponding k8s Tuned Profile resource and tuned/rendered object, which contains a list of all `TuneD` (daemon) profiles.

While this setup works for most cases, there is a problem with this approach when a cluster administator changes both a current `TuneD` profile content and (at the same) time switches to a new `TuneD` profile completely.  Then, depending on the k8s object update timing, we could see two `TuneD` daemon reloads instead of just one.

Remove the tuned/rendered object and carry `TuneD` (daemon) profiles directly in the Tuned Profile k8s objects.