@jmencak 2024-04-16
/test ci/prow/e2e-gcp-pao-workloadhints 

ci/prow/e2e-gcp-pao is a real failure, need to adjust
```
test/e2e/performanceprofile/functests/6_mustgather_testing/mustgather.go
```
to exclude the rendered resource.