@ffromani 2024-04-16
Had another pass. Can't see anything obviously wrong with the proposed approach. Need to see it running though. I'll play with this PR while reworking my #1019