@jmencak 2024-04-16
> Had another pass. Can't see anything obviously wrong with the proposed approach. Need to see it running though. I'll play with this PR while reworking my #1019

Thank you.  I'll try to address the conserns ASAP.  Please keep in mind that this might not be backportable.  I'm open to discussion about this.