@jmencak 2024-04-18
> @jmencak: trigger 5 job(s) of type blocking for the ci release of OCP 4.16
> 
>     * periodic-ci-openshift-release-master-ci-4.16-upgrade-from-stable-4.15-e2e-aws-ovn-upgrade
> 
>     * periodic-ci-openshift-release-master-ci-4.16-upgrade-from-stable-4.15-e2e-azure-sdn-upgrade
> 
>     * periodic-ci-openshift-release-master-ci-4.16-e2e-gcp-ovn-upgrade
> 
>     * periodic-ci-openshift-release-master-ci-4.16-e2e-aws-sdn-serial
> 
>     * periodic-ci-openshift-hypershift-release-4.16-periodics-e2e-aws-ovn
> 
> 
> See details on https://pr-payload-tests.ci.openshift.org/runs/ci/ddb268d0-fcf8-11ee-925a-ec8807a0928d-0

The 
```
[sig-arch][Late] operators should not create watch channels very often [apigroup:apiserver.openshift.io] [Suite:openshift/conformance/parallel]
```
test seems to be passing.  I'll also test manually and compare the number of watches with a cluster prior this PR.