@jmencak 2024-04-18
I've ran some performance testing to evaluate the impact of this PR on the CPU/memory and the number of watches.

I've tested both during 1h idle time and while running the NTO test-e2e tests.
I've tested on two SNO installs, one was a Dell PowerEdge R640 (Intel), the other one PowerEdge R6525 (AMD).

The number of watches collected via the "kubectl-dev_tool audit" command was very close.

This is an example for the Intel system:

 **With this PR**
```
count: 165, first: 2024-04-18T12:39:22+02:00, last: 2024-04-18T13:37:18+02:00, duration: 57m55.911899s
17x                  tuned.openshift.io/tuneds
16x                  machineconfiguration.openshift.io/v1/machineconfigs
16x                  tuned.openshift.io/profiles
15x                  machineconfiguration.openshift.io/v1/machineconfigpools
15x                  v1/nodes
9x                   tuned.openshift.io/v1/profiles
9x                   machineconfiguration.openshift.io/v1/kubeletconfigs
8x                   config.openshift.io/v1/clusterversions
8x                   config.openshift.io/v1/clusteroperators
8x                   tuned.openshift.io/v1/tuneds

```

**Without this PR**
```
count: 155, first: 2024-04-18T10:41:38+02:00, last: 2024-04-18T11:35:38+02:00, duration: 54m0.239677s
15x                  tuned.openshift.io/tuneds
15x                  machineconfiguration.openshift.io/v1/machineconfigs
15x                  machineconfiguration.openshift.io/v1/machineconfigpools
15x                  tuned.openshift.io/profiles
14x                  v1/nodes
8x                   tuned.openshift.io/v1/profiles
8x                   apps/daemonsets
8x                   config.openshift.io/v1/featuregates
8x                   machineconfiguration.openshift.io/v1/containerruntimeconfigs
8x                   v1/configmaps
```

The slight difference could be due to the cca 4 minute difference in duration. For the AMD system, the results were nearly exactly the same.

CPU utilization was measured both for the operator and operand in user and kernel space.

**1h during idle with this PR**
```
                user        kernel
CPU (operator): 253         82
CPU (operand):  118         29

                VmSize      VmRSS
MEM (operator): 4243056     77376
MEM (operand):  2912040     58468
```


**1h during idle without this PR**

```
                user        kernel
CPU (operator): 255         81
CPU (operand):  81          34

                VmSize      VmRSS
MEM (operator): 4612492     79440
MEM (operand):  2912304     60532

```

**While running the `make test-e2e` with this PR**
```
                user        kernel
CPU (operator): 101         16
CPU (operand):  22          13

                VmSize      VmRSS
MEM (operator): 5204580     93656
MEM (operand):  2986796     61684
```

**While running the `make test-e2e` without this PR**
```
                user        kernel
CPU (operator): 105         13
CPU (operand):  25          12

                VmSize      VmRSS
MEM (operator): 5277808     96648
MEM (operand):  3281732     60532
```

To me, the numbers look very similar with/without this PR.